Store keys on parents using Toy::Store::All
When an item is saved using Toy::Store::All store the all collection on the item and on its parent if it is also a Toy::Store::All item.  For example, in spirit there are devices and lights.  Light is a child of device.  When I save a light I should see it in the list of Light.all and Device.all since a light is a device.